Requery ontology to get parent properties
Properties that were created from generated classes were unable to be imported due to errors that the ingest repository threw because it does not pull back parent concept properties. This behavior is different between the VertexiumOntologyRepository and the InMemoryOntologyRepository. This fix can be made in the VertexiumOntologyRepository by querying the getConceptsWithProperties() while trying to pull back the parent concept, but I was split on where I wanted to make the fix.