Fix ontology repository to return parent properties (#936)
* Fix ontology repository to return parent propeties

Whenever the the vertexium ontology repository would query for a parent concept, it would not return the properties that were associated with that concept. This is important when checking inheritance in the case of the generated ontology code, and would cause the IngestRepository to start throwing failures because it tries to prevent invalid data from going into Visallo. By returning a more fully inflated parent concept the VertexiumOntologyRepository not only matches what is expected, but now works the same as in in memory implementation

* add unit tests for vertexium parent concept pulling back properties

* use parent iri from vertex to find parent concept

Conflicts:
	core/plugins/model-vertexium/src/test/java/org/visallo/vertexium/model/ontology/VertexiumOntologyRepositoryTest.java

Conflicts:
	core/plugins/model-vertexium/src/test/java/org/visallo/vertexium/model/ontology/VertexiumOntologyRepositoryTest.java